update rulesets: allow code-share action/pre-action
CMK-17294
