+++ cmk/update_config/plugins/actions/bi_config.py
@@ -9,7 +9,7 @@ from cmk.gui.bi._packs import get_cached_bi_packs
 
 from cmk.bi.packs import BIAggregationPacks
 from cmk.bi.search import BIHostSearch, BIServiceSearch
-from cmk.update_config.plugins.actions.rulesets import transform_condition_labels_to_label_groups
+from cmk.update_config.plugins.lib.rulesets import transform_condition_labels_to_label_groups
 from cmk.update_config.registry import update_action_registry, UpdateAction
